chore: locate & move `service` tests
