+++ services/Cargo.toml
@@ -25,6 +25,8 @@ getset = { workspace = true }
 json = { workspace = true }
 serde_json = { workspace = true }
 hex = { workspace = true }
+tokio = { workspace = true }
+tracing-subscriber = { workspace = true }
 
 [build-dependencies]
 hex = { workspace = true }
